// src/types/profile.ts
export interface ProfileData {
  id: string;
  first_name: string | null;
  last_name: string | null;
  avatar_url: string | null;
  phone_number: string | null;
  date_of_birth: string | null;
  gender: string | null;
  address: string | null;
  blood_type: string | null;
  emergency_contact: string | null;
  health_id: string | null;
  user_role: string | null;
  created_at: string | null;
  updated_at: string | null;
}